rename to StrategyFilter
