+++ src/JiraStatusNotifier/Jira/TicketsByAssignee/StrategyFilter/StrategyFilter.php
@@ -2,11 +2,11 @@
 
 declare(strict_types=1);
 
-namespace Chemaclass\JiraStatusNotifier\Jira\TicketsByAssignee\FilterStrategy;
+namespace Chemaclass\JiraStatusNotifier\Jira\TicketsByAssignee\StrategyFilter;
 
 use Chemaclass\JiraStatusNotifier\Jira\ReadModel\JiraTicket;
 
-interface FilterStrategy
+interface StrategyFilter
 {
     public function isValidTicket(JiraTicket $ticket): bool;
 }
