+++ src/JiraStatusNotifier/Jira/TicketsByAssignee/StrategyFilter/TicketFilter.php
@@ -1,0 +1,27 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Chemaclass\JiraStatusNotifier\Jira\TicketsByAssignee\StrategyFilter;
+
+use Chemaclass\JiraStatusNotifier\Jira\ReadModel\JiraTicket;
+
+final class TicketFilter
+{
+    private StrategyFilter $strategyFilter;
+
+    public static function notWithAssigneeKeys(string...$assigneeKeys): self
+    {
+        return new self(new NotWithAssigneeKeysStrategy(...$assigneeKeys));
+    }
+
+    private function __construct(StrategyFilter $strategyFilter)
+    {
+        $this->strategyFilter = $strategyFilter;
+    }
+
+    public function shouldIgnore(JiraTicket $ticket): bool
+    {
+        return !$this->strategyFilter->isValidTicket($ticket);
+    }
+}
